Refactor IK and bullet code (#28)
* update and add comments

* add IK script with tests

* add ik test

* issue with inverse kinematics - solution is incorrect

* update robot testing - make sure the joints are right for manip-mode ik
solving in the old robot class

* increase iterations count

* update articulated object code - why does it not reach the correct
position

* fix an issue with the q0 configuration for pybullet planner

* invert before applying

* update - cannot invert the robot pose - probably a bug

* update arm configuration for the robot

* verification for the robot joint states

* update phys cli id

* update code, reorg and cleanup for manipulation code

* update translations between robot shape and dofs

* setting initial configuration for the robot before solving

* update the robot configuration for predicting goals

* add gripper offset

* add some debug information about x motions, etc

* fix formatting issues with the IK code

* add IK test for a few hard poses

* add IK test, refactor it a bit

* update tests via black

* refactor + cleanup IK code

* black formatting

* add IK tests

* add quaternion distance (not using angle)

* update ik fixes

* Fix pathing

* Fix pathing of red block

---------
